Moved existing code for modules benchmarker, converter and launcher client to version V1.
