+++ netspresso/benchmarker/__init__.py
@@ -1,3 +1,3 @@
-from .benchmarker import Benchmarker
+from netspresso.benchmarker.v1.benchmarker import Benchmarker
 
 __all__ = ["Benchmarker"]
