+++ netspresso/clients/launcher/__init__.py
@@ -1,3 +1,3 @@
-from .main import launcher_client
+from netspresso.clients.launcher.v1.main import launcher_client
 
 __all__ = ["launcher_client"]
