+++ netspresso/converter/__init__.py
@@ -1,3 +1,3 @@
-from .converter import Converter
+from netspresso.converter.v1.converter import Converter
 
 __all__ = ["Converter"]
